HACK: Fix --apply to skip "fixup" commands related to skipped "merge".
This is not really correct as an updated or added "fixup" would not be
included.  But it's still better than amending the latest (usually
unrelated) merge in integration branch.

It would be better in that case to create an empty commit so the fixup gets
amended to that, and clean up that commit if it's still empty after fixup,
but that requries more intense surgery (eg. adding "commit" lines to the
filtered instructions, and a maybe new command to discard an empty commit).